pyproject.toml
[project]
name = "green-fashion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "make>=0.1.6.post2",
    "matplotlib>=3.10.3",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.0",
    "streamlit>=1.46.1",
    "transformers>=4.53.2",
    "torch>=2.7.1,<3.0.0",
    "pre-commit>=4.2.0",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "watchdog>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["green_fashion"]
